Remove importer filter action trait and move its logic into a detector service method
