+++ src/routes/importing-route.php
@@ -7,7 +7,7 @@ use WP_REST_Response;
 use Yoast\WP\SEO\Actions\Importing\Importing_Action_Interface;
 use Yoast\WP\SEO\Conditionals\AIOSEO_V4_Importer_Conditional;
 use Yoast\WP\SEO\Main;
-use Yoast\WP\SEO\Services\Importing\Importer_Action_Filter_Trait;
+use Yoast\WP\SEO\Services\Importing\Importable_Detector_Service;
 
 /**
  * Importing_Route class.
@@ -15,8 +15,6 @@ use Yoast\WP\SEO\Services\Importing\Importer_Action_Filter_Trait;
  * Importing route for importing from other SEO plugins.
  */
 class Importing_Route extends Abstract_Action_Route {
-
-	use Importer_Action_Filter_Trait;
 
 	/**
 	 * The import route constant.
@@ -33,12 +31,24 @@ class Importing_Route extends Abstract_Action_Route {
 	protected $importers = [];
 
 	/**
+	 * The importable detector service.
+	 *
+	 * @var Importable_Detector_Service
+	 */
+	protected $importable_detector;
+
+	/**
 	 * Importing_Route constructor.
 	 *
-	 * @param Importing_Action_Interface ...$importers All available importers.
+	 * @param Importable_Detector_Service $importable_detector The importable detector service.
+	 * @param Importing_Action_Interface  ...$importers        All available importers.
 	 */
-	public function __construct( Importing_Action_Interface ...$importers ) {
-		$this->importers = $importers;
+	public function __construct(
+		Importable_Detector_Service $importable_detector,
+		Importing_Action_Interface ...$importers
+	) {
+		$this->importable_detector = $importable_detector;
+		$this->importers           = $importers;
 	}
 
 	/**
@@ -121,7 +131,7 @@ class Importing_Route extends Abstract_Action_Route {
 	 * @return Importing_Action_Interface|false The importer, or false if no importer was found.
 	 */
 	protected function get_importer( $plugin, $type ) {
-		$importers = $this->filter_actions( $this->importers, $plugin, $type );
+		$importers = $this->importable_detector->filter_actions( $this->importers, $plugin, $type );
 
 		if ( count( $importers ) !== 1 ) {
 			return false;
